server/requirements.txt
twelvelabs==1.0.1
flask==3.0.0
python-dotenv==1.0.0
flask-cors==4.0.0
cohere==5.11.0
anthropic==0.40.0
PyGithub==2.1.1
requests==2.31.0
gitpython==3.1.40
httpx==0.27.2